pyproject.toml
[project]
name = "agente-telegram"
version = "0.1.0"
description = ""
authors = [
    {name = "Kauan Monteiro",email = "kaauan748@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytelegrambotapi (>=4.32.0,<5.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "google-genai (>=1.67.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "agente_telegram", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
